fix(mcp): add official registry metadata
